Refactor removing the jni subpackage
