+++ .gitignore
@@ -3,10 +3,10 @@
 /measurement_kit_android-v*.tar.bz2.asc
 /measurement_kit-v*-android_jni*.tar.bz2
 /measurement_kit-v*-android_jni*.tar.bz2.asc
-/jni/org_openobservatory_measurement_kit_jni_*.h
+/jni/org_openobservatory_measurement_kit_*.h
 /jni/conf
 /jni/spec/
-/java/org/openobservatory/measurement_kit/jni/OoniTestWrapper.java
-/java/org/openobservatory/measurement_kit/jni/OoniTestWrapperSwig.java
-/java/org/openobservatory/measurement_kit/jni/OoniTestWrapperSwigJNI.java
+/java/org/openobservatory/measurement_kit/OoniTestWrapper.java
+/java/org/openobservatory/measurement_kit/OoniTestWrapperSwig.java
+/java/org/openobservatory/measurement_kit/OoniTestWrapperSwigJNI.java
 /jni/ooni_test_wrapper_wrap.cpp
