+++ Makefile
@@ -12,9 +12,9 @@ BASEURL   = https://github.com/measurement-kit/measurement-kit/releases/download
 VERSION   = v0.3.0-alpha.1
 TAG       = -android_jni
 INPUT     = measurement_kit-$(VERSION)$(TAG).tar.bz2
-OVERSION  = $(VERSION)-1
+OVERSION  = $(VERSION)-2
 OUTPUT    = measurement_kit_android-$(OVERSION).tar.bz2
-PACKAGE   = org.openobservatory.measurement_kit.jni
+PACKAGE   = org.openobservatory.measurement_kit
 
 ABIS      = arm64-v8a armeabi armeabi-v7a mips mips64 x86 x86_64
