+++ java/org/openobservatory/measurement_kit/LoggerApi.java
@@ -2,7 +2,7 @@
 // Measurement-kit is free software. See AUTHORS and LICENSE for more
 // information on the copying conditions.
 
-package org.openobservatory.measurement_kit.jni;
+package org.openobservatory.measurement_kit;
 
 /**
  * Access to MeasurementKit Logger functionality
