+++ jni/portolan_sync_api.cpp
@@ -9,7 +9,7 @@
 #include <measurement_kit/traceroute.hpp>
 #include <string>
 #include "common.hpp"
-#include "org_openobservatory_measurement_kit_jni_sync_PortolanSyncApi.h"
+#include "org_openobservatory_measurement_kit_sync_PortolanSyncApi.h"
 
 using namespace mk;
 using namespace mk::traceroute;
@@ -50,7 +50,7 @@ static const char *map_code(ProbeResult r, bool timeout) {
 }
 
 JNIEXPORT jlong JNICALL
-Java_org_openobservatory_measurement_1kit_jni_sync_PortolanSyncApi_openProber
+Java_org_openobservatory_measurement_1kit_sync_PortolanSyncApi_openProber
   (JNIEnv * /*env*/, jclass /*clazz*/, jboolean use_ipv4, jint port) {
     try {
         return (jlong) new ProberContext(use_ipv4, port);
@@ -60,7 +60,7 @@ Java_org_openobservatory_measurement_1kit_jni_sync_PortolanSyncApi_openProber
 }
 
 JNIEXPORT void JNICALL
-Java_org_openobservatory_measurement_1kit_jni_sync_PortolanSyncApi_sendProbe
+Java_org_openobservatory_measurement_1kit_sync_PortolanSyncApi_sendProbe
   (JNIEnv *env, jclass /*clazz*/, jlong ptr, jstring destIp, jint destPort,
    jint ttl, jdouble timeout, jobjectArray outStrings, jintArray outInts,
    jdoubleArray outDoubles) {
@@ -120,7 +120,7 @@ Java_org_openobservatory_measurement_1kit_jni_sync_PortolanSyncApi_sendProbe
 }
 
 JNIEXPORT void JNICALL
-Java_org_openobservatory_measurement_1kit_jni_sync_PortolanSyncApi_closeProber
+Java_org_openobservatory_measurement_1kit_sync_PortolanSyncApi_closeProber
   (JNIEnv * /*env*/, jclass /*clazz*/, jlong ptr) {
     // Note: `delete` gracefully handles null pointers
     try {
@@ -132,7 +132,7 @@ Java_org_openobservatory_measurement_1kit_jni_sync_PortolanSyncApi_closeProber
 }
 
 JNIEXPORT jboolean JNICALL
-Java_org_openobservatory_measurement_1kit_jni_sync_PortolanSyncApi_checkPort
+Java_org_openobservatory_measurement_1kit_sync_PortolanSyncApi_checkPort
        (JNIEnv *env, jclass /*clazz*/, jboolean use_ipv4, jstring address,
         jstring port, jdouble timeout, jboolean verbose) {
     jboolean is_port_open = JNI_FALSE;
