+++ scripts/run-swig
@@ -2,11 +2,12 @@
 set -e
 ROOTDIR=$(cd $(dirname $(dirname $0)) && pwd -P)
 
-swig -c++ -package org.openobservatory.measurement_kit.jni                     \
+swig -c++ -package org.openobservatory.measurement_kit                         \
         -java jni/ooni_test_wrapper.i
 
-for suffix in "" Swig SwigJNI; do
+for suffix in "" SwigJNI; do
     mv jni/OoniTestWrapper${suffix}.java                                       \
-            java/org/openobservatory/measurement_kit/jni
+            java/org/openobservatory/measurement_kit
 done
+rm jni/OoniTestWrapperSwig.java  # basically an empty file
 mv jni/ooni_test_wrapper_wrap.cxx jni/ooni_test_wrapper_wrap.cpp
